% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsget.R
\name{fetch_file}
\alias{fetch_file}
\title{Title}
\usage{
fetch_file(file_id, url, token = NULL, destination)
}
\arguments{
\item{file_id}{}

\item{url}{}

\item{token}{}

\item{destination}{}
}
\value{

}
\description{
Title
}
