% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsget.R
\name{htsget_reads}
\alias{htsget_reads}
\alias{htsget_variants}
\title{Retrieve reads (BAM files) or variants (VCF files)}
\usage{
htsget_reads(granges, sample_id, url, destination)

htsget_variants(granges, sample_id, url, destination)
}
\arguments{
\item{granges}{\code{GRanges()} instance with genomic coordinates for
retrieval. Coordinates follow Bioconductor standards (1-based,
closed intervals) and are translated to GA4GH coordinates
(0-based, half-open intervals).}

\item{sample_id}{\code{character(1)} sample identifier}

\item{url}{\code{character(1)} data resource url, excluding endpoint and
sample information.}

\item{destination}{\code{character(1)} file for output; must not exist.}
}
\description{
Retrieve reads (BAM files) or variants (VCF files)
}
\examples{
gr <- GenomicRanges::GRanges("chr12:111766922-111817529")
sample_id <- "platinum/NA12878"
url <- "https://htsnexus.rnd.dnanex.us/v1"
fl <- tempfile(fileext=".bam")
bam <- htsget_reads(gr, sample_id, url, fl)
Rsamtools::countBam(bam)

gr <- GenomicRanges::GRanges("12:112204691-112247789")
sample_id <- "1000genomes/20130502_autosomes"
url <- "https://htsnexus.rnd.dnanex.us/v1"
fl <- tempfile(fileext=".vcf")
vcf <- htsget_variants(gr, sample_id, url, fl)
VariantAnnotation::readVcf(vcf)
}
