% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htsget.R
\name{htsget_reads}
\alias{htsget_reads}
\title{Retrieve reads (BAM files)}
\usage{
htsget_reads(granges, sample_id, endpoint, destination)
}
\arguments{
\item{granges}{`GRanges()` instance with genomic coordinates for
retrieval. Coordinates follow Bioconductor standards (1-based,
closed intervals) and are translated to GA4GH coordinates
(0-based, half-open intervals).}

\item{sample_id}{`character(1)` sample identifier}

\item{endpoint}{`character(1)` data resource url, excluding
endpoint.}

\item{destination}{`character(1)` file for output; must not exist.}
}
\description{
Retrieve reads (BAM files)
}
\examples{
gr <- GRanges("chr12:111766922-111817529")
sample_id <- "platinum/NA12878"
endpoint <- "https://htsnexus.rnd.dnanex.us/v1"
fl <- file.path(tempdir(), paste0(basename(url), ".bam"))
unlink(fl)
bam <- htsget_reads(gr, sample_id, endpoint, fl)
Rsamtools::countBam(bam)

}
